requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
workalendar==17.0.0
python-dateutil==2.9.0